EVNT-35 Refactor the code in the interceptors to a utility class
