+++ api-1.x/src/main/java/org/openmrs/event/api/db/hibernate/HibernateEventInterceptor_1_x.java~
@@ -1,12 +1,14 @@
 package org.openmrs.event.api.db.hibernate;
 
+import java.io.Serializable;
+import java.util.HashSet;
+import java.util.Stack;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.openmrs.annotation.OpenmrsProfile;
 
 import org.openmrs.event.HibernateEventInterceptorSuperClass;
-
 /**
  * A hibernate {@link Interceptor} implementation, intercepts any database inserts, updates and
  * deletes in a single hibernate session and fires the necessary events. Any changes/inserts/deletes
@@ -16,6 +18,7 @@ import org.openmrs.event.HibernateEventInterceptorSuperClass;
  */
 @OpenmrsProfile(openmrsPlatformVersion = "1.*")
 public class HibernateEventInterceptor_1_x extends HibernateEventInterceptorSuperClass {
+	
 	protected final Log log = LogFactory.getLog(HibernateEventInterceptor_1_x.class);
 
 }
