+++ api-2.x/pom.xml~
@@ -7,10 +7,14 @@
 		<version>2.6-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>event-api</artifactId>
+	<artifactId>event-api-2.x</artifactId>
 	<packaging>jar</packaging>
-	<name>Event Module API</name>
-	<description>API project for EventQueue</description>
+	<name>Event Module API 2.x</name>
+	<description>2.x API project for EventQueue</description>
+
+	<properties>
+        	<openMRSVersion>2.0.0-alpha</openMRSVersion>
+        </properties>
 
 	<dependencies>
 	
@@ -22,6 +26,12 @@
 		</dependency> 
 		-->
 		
+			<dependency>
+				<groupId>${project.parent.groupId}</groupId>
+				<artifactId>${project.parent.artifactId}-api</artifactId>
+				<version>${project.parent.version}</version>
+			</dependency>
+
 			<dependency>
 				<groupId>org.apache.activemq</groupId>
 				<artifactId>activemq-core</artifactId>
@@ -112,21 +122,20 @@
 						<artifactId>spring-oxm</artifactId>
 					</exclusion>
 				</exclusions>
-			</dependency>	
-
-			<!-- https://mvnrepository.com/artifact/org.hibernate/hibernate-core -->
+			</dependency>
+	
 			<dependency>
-    				<groupId>org.hibernate</groupId>
+   			 	<groupId>org.hibernate</groupId>
     				<artifactId>hibernate-core</artifactId>
     				<version>4.1.0.Final</version>
 			</dependency>
-
 
 		<!-- Begin OpenMRS core -->
 		
 		<dependency>
 			<groupId>org.openmrs.api</groupId>
 			<artifactId>openmrs-api</artifactId>
+			<version>${openMRSVersion}</version>
 			<type>jar</type>
 			<scope>provided</scope>
 		</dependency>
@@ -134,6 +143,7 @@
 		<dependency>
 			<groupId>org.openmrs.web</groupId>
 			<artifactId>openmrs-web</artifactId>
+			<version>${openMRSVersion}</version>
 			<type>jar</type>
 			<scope>provided</scope>
 		</dependency>
