+++ api-2.x/src/main/java/org/openmrs/event/api/db/hibernate/HibernateEventInterceptor_2_x.java~
@@ -3,8 +3,6 @@ package org.openmrs.event.api.db.hibernate;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.openmrs.annotation.OpenmrsProfile;
-
 import org.openmrs.event.HibernateEventInterceptorSuperClass;
 
 /**
@@ -14,8 +12,8 @@ import org.openmrs.event.HibernateEventInterceptorSuperClass;
  *
  * We use a Stack here to handle any nested transactions that may occur within a single thread
  */
-@OpenmrsProfile(openmrsPlatformVersion = "1.*")
-public class HibernateEventInterceptor_1_x extends HibernateEventInterceptorSuperClass {
-	protected final Log log = LogFactory.getLog(HibernateEventInterceptor_1_x.class);
-
+@OpenmrsProfile(openmrsPlatformVersion = "2.*")
+public class HibernateEventInterceptor_2_x extends HibernateEventInterceptorSuperClass {
+	protected final Log log = LogFactory.getLog(HibernateEventInterceptor_2_x.class);
+	
 }
